# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.example.com"
        },
        {
            "url": "https://migright.envast.tn/api/",
            "description": "Migright Staging Server"
        },
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Hr local Server"
        }
    ],
    "paths": {
...
        "/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get paginated list of news",
                "operationId": "b83ff020eb45141b07c365e50a4b27bd",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "The locale for the news items",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paginated list of news",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "perPage": {
                                            "type": "integer"
                                        },
                                        "nextCursor": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "previousCursor": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "news": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/NewsResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/news/{id}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get news by ID",
                "description": "Retrieve a news item by its ID",
                "operationId": "52293d29070c8525e0db8f2301add0b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the news item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "The locale for the news item",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The news item",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "news": {
                                            "$ref": "#/components/schemas/NewsResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "News item not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NewsResource": {
                "properties": {
                    "id": {
                        "description": "ID of the news item",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Title of the news item",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the news item",
                        "type": "string"
                    },
                    "image": {
                        "description": "Image URL of the news item",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}